+++ webpack-plugin/generate-classes.js
@@ -5,14 +5,14 @@ const generateClass = require('../lib');
 const attemptGenerateClass = ({ baseClass, modulePath, namespace, indent }) => {
   try {
     const sourceCode = fs.readFileSync(modulePath, 'utf-8');
-    const { code, componentName } = generateClass({
+    const { code, className } = generateClass({
       baseClass,
       indent,
       namespace,
       sourceCode
     });
 
-    return { code, componentName };
+    return { code, className };
   } catch (error) {
     return {
       error: `\n${modulePath}\n${error.message}\n`
@@ -25,16 +25,16 @@ const generateClasses = ({ baseClass, indent, modulePaths, namespace }) => {
   const classes = modulePaths.map(modulePath =>
     attemptGenerateClass({ baseClass, indent, modulePath, namespace })
   );
-  const duplicates = classes.reduce((accum, { componentName }, index) => {
-    if (componentName) {
+  const duplicates = classes.reduce((accum, { className }, index) => {
+    if (className) {
       const duplicateIndex = classes
         .slice(index + 1)
-        .findIndex(c => c.componentName === componentName);
+        .findIndex(c => c.className === className);
 
       if (duplicateIndex !== -1) {
         return accum.concat(
-          `${componentName} (${modulePaths[index]})`,
-          `${componentName} (${modulePaths[duplicateIndex + 1]})`
+          `${className} (${modulePaths[index]})`,
+          `${className} (${modulePaths[duplicateIndex + 1]})`
         );
       }
     }