+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Breaking: `Array` in propTypesMeta is no longer supported. Use array literals instead
 - Breaking: Node.js API now exports an object instead of a function.
+- Breaking: Node.js generator now returns `className` instead of `componentName`
 
 # 0.6.5